# openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
...
    "termsOfService" : "https://www.pagopa.gov.it/",
    "title" : "WISP Converter",
    "version" : "0.4.25"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/receipt" : {
      "get" : {
        "operationId" : "receiptRetrieve",
        "parameters" : [ {
          "in" : "query",
          "name" : "ci",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "ccp",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "iuv",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Receipt exists",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "tags" : [ "Receipt" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ]
    },
    "/receipt/ko" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "operationId" : "receiptKo",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully forwarded negative paaInviaRT to EC",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "tags" : [ "Receipt" ]
      }
    },
    "/receipt/ok" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "operationId" : "receiptOk",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReceiptRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully forwarded positive paaInviaRT to EC",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "tags" : [ "Receipt" ]
      }
    },
    "/receipt/timer" : {
      "delete" : {
        "description" : "Delete a timer by paymentToken",
        "operationId" : "deleteTimer_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "paymentTokens",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully paymentToken expiration timer deleted",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "deleteTimer",
        "tags" : [ "ReceiptTimer" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Create a timer linked with paymentToken and receipt data",
        "operationId" : "createTimer_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReceiptTimerRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully paymentToken expiration timer created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "createTimer",
        "tags" : [ "ReceiptTimer" ]
      }
    },
...
    "/recovery/receipts" : {
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Execute reconciliation of all receipts in the request, searching by passed identifier",
        "operationId" : "recoverReceiptToBeReSent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecoveryReceiptRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecoveryReceiptReportResponse"
                }
              }
            },
            "description" : "Reconciliation scheduled",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "Execute reconciliation for passed receipts.",
        "tags" : [ "Recovery" ]
      }
    },
...
    "/rpt/timer" : {
      "delete" : {
        "description" : "Delete a timer by sessionId",
        "operationId" : "deleteTimer",
        "parameters" : [ {
          "in" : "query",
          "name" : "sessionId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully rpt timer deleted",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "deleteRPTTimer",
        "tags" : [ "RPTTimer" ]
      },
      "parameters" : [ {
        "description" : "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
        "in" : "header",
        "name" : "X-Request-Id",
        "schema" : {
          "type" : "string"
        }
      } ],
      "post" : {
        "description" : "Create a timer from sessionId data",
        "operationId" : "createTimer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RPTTimerRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully rpt timer created",
            "headers" : {
              "X-Request-Id" : {
                "description" : "This header identifies the call",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "ApiKey" : [ ]
        } ],
        "summary" : "createRPTTimer",
        "tags" : [ "RPTTimer" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "PairStringString" : {
        "type" : "object",
        "properties" : {
          "first" : {
            "type" : "string"
          },
          "second" : {
            "type" : "string"
          }
        }
      },
      "RPTTimerRequest" : {
        "type" : "object",
        "properties" : {
          "sessionId" : {
            "type" : "string"
          }
        }
      },
      "ReceiptRequest" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string"
          }
        }
      },
      "ReceiptTimerRequest" : {
        "type" : "object",
        "properties" : {
          "expirationTime" : {
            "type" : "integer",
            "format" : "int64"
          },
          "fiscalCode" : {
            "pattern" : "\\w*",
            "type" : "string"
          },
          "noticeNumber" : {
            "pattern" : "\\d*",
            "type" : "string"
          },
          "paymentToken" : {
            "type" : "string"
          },
          "sessionId" : {
            "type" : "string"
          }
        }
      },
...
      "RecoveryReceiptReportResponse" : {
        "type" : "object",
        "properties" : {
          "receiptStatus" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PairStringString"
            }
          }
        }
      },
      "RecoveryReceiptRequest" : {
        "type" : "object",
        "properties" : {
          "receiptIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "ApiKey" : {
        "description" : "The API key to access this function app.",
        "in" : "header",
        "name" : "Ocp-Apim-Subscription-Key",
        "type" : "apiKey"
      }
    }
  }
}